Gmr hack (#29)
Functional VCO-1 release

+++ dsp/fft/FFT.cpp
@@ -82,16 +82,16 @@ bool FFT::inverse(FFTDataReal* out, const FFTDataCpx& in)
     return true;
 }
 
-int FFT::freqToBin(float freq, float sampleRate, int numBins)
+int FFT::freqToBin(double freq, double sampleRate, int numBins)
 {
     assert(freq <= (sampleRate / 2));
     // bin(numBins) <> sr / 2;
     return (int)((freq / sampleRate)*(numBins));
 }
 
-float FFT::bin2Freq(int bin, float sampleRate, int numBins)
+double FFT::bin2Freq(int bin, double sampleRate, int numBins)
 {
-    return  sampleRate * float(bin) / (float(numBins));
+    return  sampleRate * double(bin) / double(numBins);
 }
 
 static float randomPhase()
@@ -119,8 +119,8 @@ static void makeNegSlope(FFTDataCpx* output, const ColoredNoiseSpec& spec)
     static float k = -spec.slope * log2(lowFreqCorner);
     for (int i = bin40 + 1; i < numBins; ++i) {
         if (i < numBins / 2) {
-            const float f = FFT::bin2Freq(i, spec.sampleRate, numBins);
-            const float gainDb = std::log2(f) * spec.slope + k;
+            const double f = FFT::bin2Freq(i, spec.sampleRate, numBins);
+            const double gainDb = std::log2(f) * spec.slope + k;
             const float gain = float(AudioMath::gainFromDb(gainDb));
             output->set(i, std::polar(gain, randomPhase()));
         } else {
@@ -142,8 +142,8 @@ static void makePosSlope(FFTDataCpx* output, const ColoredNoiseSpec& spec)
     static float k = -spec.slope * log2(spec.highFreqCorner);
     for (int i = binHigh - 1; i > 0; --i) {
         if (i < numBins / 2) {
-            const float f = FFT::bin2Freq(i, spec.sampleRate, numBins);
-            const float gainDb = std::log2(f) * spec.slope + k;
+            const double f = FFT::bin2Freq(i, spec.sampleRate, numBins);
+            const double gainDb = std::log2(f) * spec.slope + k;
             const float gain = float(AudioMath::gainFromDb(gainDb));
             gainMax = std::max(gain, gainMax);
             output->set(i, std::polar(gain, randomPhase()));
